chore(TSP-1): move files into proper directories
