--- src/Types.tsx
@@ -1,24 +1,0 @@
-export type Location = {
-    lat: number;
-    lon: number;
-    display_name: string
-}
-
-export type Result = {
-    distances: [[
-        {
-            distance: number;
-            originNode: Location;
-            targetNode: Location;
-            time: number;
-        }
-    ]],
-    routes: [[]]
-}
-
-export type Edge = {
-    distance: number;
-    originNode: Location;
-    targetNode: Location;
-    time: number;
-}
