+++ src/api/usePostNodes.ts
@@ -1,6 +1,6 @@
 import { useQuery } from "@tanstack/react-query";
 import axios from "axios";
-import type { Location } from "../LocationContext";
+import type { Location } from "../util/LocationContext";
 import nodes from "../../nodes.json";
 
 type UsePostNodesParameters = {
