+++ src/components/EdgesResults.tsx
@@ -9,11 +9,14 @@ import {
 	TableRow,
 } from "@mui/material";
 import { useContext, useEffect } from "react";
-import { LocationContext, type LocationContextType } from "./LocationContext";
-import { usePostNodes } from "./api/usePostNodes";
+import {
+	LocationContext,
+	type LocationContextType,
+} from "../util/LocationContext";
+import { usePostNodes } from "../api/usePostNodes";
 import { CalculateButton } from "./CalculateButton";
-import { usePostRoutes } from "./api/usePostRoutes";
-import { NodeContext, type NodeContextType } from "./NodeContext";
+import { usePostRoutes } from "../api/usePostRoutes";
+import { NodeContext, type NodeContextType } from "../util/NodeContext";
 
 type RowType = {
 	targetDisplayName: string;
