+++ src/components/NominatimNodeList.tsx
